+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes
 added;
 - Function `trmplotminatomdist` to plot minimal interatomic distances; 
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
+- Option to consider planar angles added to `trmdistantangleindices`;
 - `trmcreate` creates a alpha carbon-based transformation model if the
 specified PDB structures contains only alpha carbon atoms;
 - `trmplottranglediff` replaced with the `trmplotanglediff` function that